Finish `splitgraph.engine` cleanup
Move everything else (apart from ResultShape) into splitgraph.core.migration
