+++ test/splitgraph/test_security.py
@@ -6,9 +6,9 @@ from psycopg2._psycopg import ProgrammingError
 from psycopg2.sql import SQL, Identifier
 
 from splitgraph.config import SPLITGRAPH_META_SCHEMA
+from splitgraph.core.migration import META_TABLES
 from splitgraph.core.repository import Repository, clone
 from splitgraph.core.sql.queries import select
-from splitgraph.engine import META_TABLES
 
 
 @pytest.mark.registry
